Run local dnsmasq
Each of our microservices typically have their own dnsmasq. This is used
with serf in order to resolve the names of each microservice. We also
have a serpate microservice called dnsmasq which is for microservices
which cannot use their local dnsmasq for reasons (see proxy commit
logs).

By the looks of it mongo does not have a valid reason to not use the
local dnsmasq.